# [pr] move has_valid into pm_lowerer