Resolves `TODO: check has_valid in UPat, not here`