@geohot 2024-12-18
How does speed compare?