@leopf 2024-12-18
Ran `test/external/external_benchmark_schedule.py` on my machine 3 times each.

master:
`model lower in      25.38 ms`
`model lower in      25.54 ms`
`model lower in      25.32 ms`

after:
`model lower in      27.29 ms`
`model lower in      27.74 ms`
`model lower in      26.98 ms`